SmartHighwayFunctionApp/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions
numpy
opencv-python
cmake
dlib
azure-cosmos
signalrcore
matplotlib
azure-storage
azure-storage-blob